requirements.txt
boto3==1.28.44
pdf2image==1.16.3
Pillow==9.5.0
pytesseract==0.3.10
python_docx==0.8.11
streamlit==1.26.0
python-dotenv